% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibrarySPoTApi.r
\name{LibrarySPoTApi_add_libraries}
\alias{LibrarySPoTApi_add_libraries}
\title{Add new library object}
\usage{
LibrarySPoTApi_add_libraries(source, ...)
}
\arguments{
\item{source}{}
}
\description{
Add new library object
}
