% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibrarySPoTApi.r
\name{LibrarySPoTApi_delete_library}
\alias{LibrarySPoTApi_delete_library}
\title{Delete a library object}
\usage{
LibrarySPoTApi_delete_library(id, ...)
}
\arguments{
\item{id}{Unique identifier (accession) of the object.}
}
\description{
Delete a library object
}
