% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibrarySPoTApi.r
\name{LibrarySPoTApi_delete_library_group}
\alias{LibrarySPoTApi_delete_library_group}
\title{Delete a library group with all related library objects}
\usage{
LibrarySPoTApi_delete_library_group(id, ...)
}
\arguments{
\item{id}{Unique identifier (accession) of the object.}
}
\description{
Delete a library group with all related library objects
}
