% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibrarySPoTApi.r
\name{LibrarySPoTApi_get_libraries_by_group}
\alias{LibrarySPoTApi_get_libraries_by_group}
\title{Retrieve libraries related to the given group}
\usage{
LibrarySPoTApi_get_libraries_by_group(
  id,
  returned_metadata_fields,
  page_offset,
  page_limit,
  ...
)
}
\arguments{
\item{id}{Unique identifier (accession) of the object.}

\item{returned_metadata_fields}{The parameter defines amount of metadata attributes to return:  1. \code{template} - return metadata attributes according to applied template, if object doesn’t have applied template default template will be used. This is the default for User endpoints. 2. \code{all} - return all metadata attributes with values and null attributes, if they are present in the applied template. This is the default for Curator endpoints.}

\item{page_offset}{Show the page {pageOffset+1} results from the start of the results. E.g. 100 will show a page of results  starting from the 101st result. The default value is 0.}

\item{page_limit}{Maximum number of results to return per page (see Paging above). This value must be between 0 and 2000 (inclusive). The default is 2000.}
}
\description{
Retrieve libraries related to the given group
}
