% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibrarySPoTApi.r
\name{LibrarySPoTApi_update_library}
\alias{LibrarySPoTApi_update_library}
\title{Update a library object}
\usage{
LibrarySPoTApi_update_library(id, body, ...)
}
\arguments{
\item{id}{Unique identifier (accession) of the object.}

\item{body}{Metadata in the form of \code{{key: value, key2: value2, ...}}}
}
\description{
Update a library object
}
